backend/requirements.txt
# 核心框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# 数据库 & 缓存
sqlalchemy==2.0.25
asyncpg==0.29.0
redis==5.0.1

# AI & 爬虫
langchain==0.1.0
langchain-openai==0.0.2
# 如果用 Gemini:
langchain-google-genai
playwright==1.41.0
beautifulsoup4==4.12.3

# 数据处理
pandas==2.2.0
numpy==1.26.3
snownlp==0.12.3

# 工具
python-dotenv==1.0.1
loguru==0.7.2
apscheduler==3.10.4